+++ tests/pdf/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for the PDF export of both the :mod:`integreat_cms.cms` and the :mod:`integreat_cms.api` app
+"""