+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 UNRELEASED
 ----------
 
+
+2022.4.1
+--------
+
 * [ [#1354](https://github.com/digitalfabrik/integreat-cms/issues/1354) ] Fix order of root pages
 * [ [#1353](https://github.com/digitalfabrik/integreat-cms/issues/1353) ] Add tunews setting to region model
 * [ [#1328](https://github.com/digitalfabrik/integreat-cms/issues/1328) ] Fix missing entries in broken link checker